moved find_executable from galgebra/latex_ex.py to utilities/misc.py
